Carruseles de vistas individuales de phones y tvs tanto para imagenes de producto en grande como para imagenes de productos relacionados, cambios de estilos en cards de carrusel
