+++ src/front/js/pages/vistaIndividualPhone.js
@@ -1,185 +1,30 @@
 import React, { useState, useContext, useEffect } from "react";
 import { Context } from "../store/appContext";
-import { ProductColors } from "../component/product-colors";
 import { RelatedProducts } from "../component/related-products";
-import { useParams } from "react-router-dom";
-import { BackendURL } from "../component/backendURL";
+import { ProductInfoPhone } from "../component/productInfoPhone";
 
-export const VistaIndividualPhone = ({product}) => {
-    const [imageColors, setImageColors] = useState("https://images.samsung.com/es/smartphones/galaxy-s25-ultra/images/galaxy-s25-ultra-features-ecosystem-galaxy-s25-mo.jpg?imbypass=true");
+export const VistaIndividualPhone = () => {
 
-    const [phone, setPhone] = useState([]);
-
-    const { store, actions } = useContext(Context)
-
-    const { smartphone_id } = useParams();
-
-    const precio = parseInt(phone.precio);
-    const totalPrecioEur = new Intl.NumberFormat("de-DE", {
-        style: "currency",
-        currency: "EUR",
-    }).format(precio);
-
-    const handleImageColors = (imageUrl) => {
-        setImageColors(imageUrl);
-    };
-
-    const getPhoneById = async () => {
-        const urlBackend = process.env.BACKEND_URL
-        try {
-            const response = await fetch(urlBackend + "phone/" + smartphone_id);
-
-            const data = await response.json();
-
-            setPhone(data);
-        } catch (error) {
-            console.error("Error getting ID phones from API");
-        }
-    };
-
-    useEffect(() => {
-        getPhoneById()
-    }, []);
+    const { actions } = useContext(Context)
 
     useEffect(() => {
         actions.setNegativeColors()
-        actions.setNavbarVisibility()
     }, [])
 
+    useEffect(() => {
+
+        document.body.style.backgroundColor = "rgb(234, 248, 252)";
+
+        return () => {
+
+            document.body.style.backgroundColor = "";
+        };
+    }, []);
+
     return (
-        <div className="container">
-            <div className="row col-md-12">
-                <div className="row col-md-5 card-individual-image">
-                    <img className="mb-5" src={imageColors} alt="Product" />
-                </div>
-                <div className="col-md-7 text-white">
-                    <div className="card-body-individual">
-                        <h5 className="card-title sub-titulo mb-3 titulo">{phone?.modelo}</h5>
-                        <p className="card-text-score"></p>
-                        <p className="card-text-price titulo">
-                            <small className="text-white">{totalPrecioEur}</small>
-                        </p>
-                        <p className="card-text">
-                            <button className="btn-add-cart texto">
-                                <i className="fa-solid fa-cart-plus mb-1"></i> Añadir al carrito
-                            </button>
-                        </p>
-                        <div className="accordion" id="accordionExample">
-                            <div className="accordion-item">
-                                <h2 className="accordion-header">
-                                    <button className="accordion-button collapsed" type="button" data-bs-toggle="collapse" data-bs-target="#collapseTop" aria-expanded="false" aria-controls="collapseTop">
-                                        <strong className="texto">Descripción</strong>
-                                    </button>
-                                </h2>
-                                <div id="collapseTop" className="accordion-collapse collapse" data-bs-parent="#accordionExample">
-                                    <div className="accordion-body">
-                                        <p className="texto">{phone?.descripcion}</p>
-                                    </div>
-                                </div>
-                            </div>
-                            <div className="accordion-item">
-                                <h2 className="accordion-header">
-                                    <button className="accordion-button collapsed" type="button" data-bs-toggle="collapse" data-bs-target="#collapseOne" aria-expanded="false" aria-controls="collapseOne">
-                                        <strong className="texto">Colores</strong>
-                                    </button>
-                                </h2>
-                                <div id="collapseOne" className="accordion-collapse collapse" data-bs-parent="#accordionExample">
-                                    <div className="accordion-body">
-                                        <div className="row ms-2">
-                                            <div className="col-md-4 d-flex flex-column align-items-center">
-                                                <h6 className="title-color">Azul</h6>
-                                                <ProductColors image="https://images.samsung.com/es/smartphones/galaxy-s25-ultra/images/galaxy-s25-ultra-features-ecosystem-galaxy-s25-mo.jpg?imbypass=true" />
-                                            </div>
-                                            <div className="col-md-4 d-flex flex-column align-items-center">
-                                                <h6 className="title-color">Negro</h6>
-                                                <ProductColors image="https://img.pccomponentes.com/articles/1086/10866751/1707-samsung-galaxy-s25-ultra-smartphone-con-ia-almacenamiento-512gb-bateria-5000mah-titanio-negro-comprar.jpg" />
-                                            </div>
-                                            <div className="col-md-4 d-flex flex-column align-items-center">
-                                                <h6 className="title-color">Plata</h6>
-                                                <ProductColors image="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRFaO3kc8c5nLkQ2FitXAbl0nGa4rS_dd8zuw&s" />
-                                            </div>
-                                        </div>
-                                    </div>
-                                </div>
-                            </div>
-                            <div className="accordion-item">
-                                <h2 className="accordion-header">
-                                    <button className="accordion-button collapsed" type="button" data-bs-toggle="collapse" data-bs-target="#collapseTwo" aria-expanded="false" aria-controls="collapseTwo">
-                                        <strong className="texto">Cámara</strong>
-                                    </button>
-                                </h2>
-                                <div id="collapseTwo" className="accordion-collapse collapse" data-bs-parent="#accordionExample">
-                                    <div className="accordion-body">
-                                        <p className="texto">{phone?.camara}</p>
-                                    </div>
-                                </div>
-                            </div>
-                            <div className="accordion-item">
-                                <h2 className="accordion-header">
-                                    <button className="accordion-button collapsed" type="button" data-bs-toggle="collapse" data-bs-target="#collapseThree" aria-expanded="false" aria-controls="collapseThree">
-                                        <strong className="texto">Almacenamiento</strong>
-                                    </button>
-                                </h2>
-                                <div id="collapseThree" className="accordion-collapse collapse" data-bs-parent="#accordionExample">
-                                    <div className="accordion-body">
-                                        <p className="texto">{phone?.almacenamiento}</p>
-                                    </div>
-                                </div>
-                            </div>
-                            <div className="accordion-item">
-                                <h2 className="accordion-header">
-                                    <button className="accordion-button collapsed" type="button" data-bs-toggle="collapse" data-bs-target="#collapseFour" aria-expanded="false" aria-controls="collapseFour">
-                                        <strong className="texto">Memoria RAM</strong>
-                                    </button>
-                                </h2>
-                                <div id="collapseFour" className="accordion-collapse collapse" data-bs-parent="#accordionExample">
-                                    <div className="accordion-body">
-                                        <p className="texto">{phone?.memoria_ram}</p>
-                                    </div>
-                                </div>
-                            </div>
-                            <div className="accordion-item">
-                                <h2 className="accordion-header">
-                                    <button className="accordion-button collapsed" type="button" data-bs-toggle="collapse" data-bs-target="#collapseFive" aria-expanded="false" aria-controls="collapseFive">
-                                        <strong className="texto">Pantalla</strong>
-                                    </button>
-                                </h2>
-                                <div id="collapseFive" className="accordion-collapse collapse" data-bs-parent="#accordionExample">
-                                    <div className="accordion-body">
-                                        <p className="texto">{phone?.pantalla}</p>
-                                    </div>
-                                </div>
-                            </div>
-                            <div className="accordion-item">
-                                <h2 className="accordion-header">
-                                    <button className="accordion-button collapsed" type="button" data-bs-toggle="collapse" data-bs-target="#collapseSix" aria-expanded="false" aria-controls="collapseSix">
-                                        <strong className="texto">Batería</strong>
-                                    </button>
-                                </h2>
-                                <div id="collapseSix" className="accordion-collapse collapse" data-bs-parent="#accordionExample">
-                                    <div className="accordion-body">
-                                        <p className="texto">{phone?.bateria}</p>
-                                    </div>
-                                </div>
-                            </div>
-                            <div className="accordion-item">
-                                <h2 className="accordion-header">
-                                    <button className="accordion-button collapsed" type="button" data-bs-toggle="collapse" data-bs-target="#collapseSeven" aria-expanded="false" aria-controls="collapseSeven">
-                                        <strong className="texto">Procesador</strong>
-                                    </button>
-                                </h2>
-                                <div id="collapseSeven" className="accordion-collapse collapse" data-bs-parent="#accordionExample">
-                                    <div className="accordion-body">
-                                        <p className="texto">{phone?.procesador}</p>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <h1 className="text-center text-white texto related-products">Productos relacionados (Smartphones)</h1>
-            <RelatedProducts />
-        </div>
+        <>
+            <ProductInfoPhone />
+            <RelatedProducts productType="phones" />
+        </>
     );
 };
