+++ src/front/styles/index.css
@@ -108,12 +108,14 @@ ul {
     transform: translate(3px, 3px);
 }
 
-.dropdown, .dropdown-search {
+.dropdown,
+.dropdown-search {
     position: relative;
     display: inline-block;
 }
 
-.dropdown-content, .dropdown-content-search{
+.dropdown-content,
+.dropdown-content-search {
     display: none;
     position: absolute;
     background-color: #f1f1f1;
@@ -122,20 +124,21 @@ ul {
     z-index: 1;
 }
 
-.dropdown-content-search{
+.dropdown-content-search {
     overflow-y: auto;
     max-height: 15em;
     width: 20em;
 }
 
-.dropdown-content a, .dropdown-content-search a  {
+.dropdown-content a,
+.dropdown-content-search a {
     color: black;
     padding: 5px 0;
     text-decoration: none;
     display: block;
 }
 
-.dropdown:hover .dropdown-content{
+.dropdown:hover .dropdown-content {
     display: block;
 }
 
@@ -422,9 +425,9 @@ ul {
 }
 
 .card-individual-tv-image img {
-    width: 800px;
-    height: 600px;
-    margin-top: 110px;
+    width: 510px;
+    height: 500px;
+    margin-top: 130px;
     object-fit: cover;
     border-radius: 8%;
     transition: transform 0.3s ease-in-out;
@@ -456,7 +459,7 @@ ul {
     color: #2F414F;
 }
 
-.container_card_catalog{
+.container_card_catalog {
     width: 30em;
 }
 
@@ -514,7 +517,7 @@ ul {
     margin: 1.5em 1em 1em;
 }
 
-.btn_card_product{
+.btn_card_product {
     background-color: rgb(215, 44, 32);
     width: 15em;
     height: 3em;
@@ -525,10 +528,12 @@ ul {
     color: #fafafa;
     font-weight: bold;
     font-size: 1em;
-};
+}
+
+;
 
 
-.negative-background{
+.negative-background {
     background-color: rgb(234, 248, 252)
 }
 
@@ -551,20 +556,29 @@ ul {
 }
 
 .cart-delete-button:hover {
-background-color: rgb(221, 221, 221);
-border: 0.01em solid black;
+    background-color: rgb(221, 221, 221);
+    border: 0.01em solid black;
 
 }
-
 
 .card-cart-product {
     background-color: white;
 }
 
-.title-cart-product {
-    font-size: 1.3em;
-    font-weight: 500;
-    margin-bottom: 0;
+.title-color {
+    margin-bottom: 10px;
+}
+
+.img-small-vista-individual {
+    margin-top: 0;
+    width: 50px;
+    height: 50px;
+    border-radius: 8%;
+}
+
+.img-small-vista-individual:hover {
+    transform: scale(1.25);
+    transition: transform 0.5s ease;
 }
 
 .description-cart-product {
@@ -574,7 +588,7 @@ border: 0.01em solid black;
 }
 
 .order-resume {
-    margin-top: 5em!important;
+    margin-top: 5em !important;
 }
 
 .order-resume-table {
@@ -588,8 +602,8 @@ border: 0.01em solid black;
 .sub-total-resume {
     font-size: 3.5em;
     font-weight: 500;
-    top:0;
-    
+    top: 0;
+
 }
 
 .resume-button {
